+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 2.0.1 - 2020-06-01
+### Fixed
+- Fixes an issue where it could be possible save multiple conditional records per field layout
+- Fixes an issue where it wasn't possible to clear conditionals
+
 ## 2.0.0 - 2020-06-01
 ### Added
 - Initial public release